Refactor focus management system: Introduced Pydantic models for data validation, implemented focus analysis and monitoring services, and organized application structure into models, services, and controllers. Removed deprecated configuration and main files, ensuring a cleaner architecture and improved maintainability.
