+++ src/__init__.py
@@ -1,1 +1,6 @@
 # src package
+# TODO: it is better to make all imports goes like the below and src should not be regarded as the root of the package name 
+
+# exmpale
+# import app.apple_model
+# instead of src.app.apple_model.
