+++ src/app/controllers/__init__.py
@@ -1,0 +1,1 @@
+# controllers package
