+++ src/app/models/__init__.py
@@ -1,0 +1,3 @@
+# models package
+# TODO: i think pydantic is needed.
+from pydantic import BaseModel, Field
