--- src/controllers/__init__.py
@@ -1,1 +1,0 @@
-# controllers package
